[RC] add regular

+++ fcn/base.py
@@ -633,7 +633,7 @@ class NN:
         with tf.name_scope('regularize'):
             regularizer = 0.0
             for var in trainable_var:
-                regularizer = tf.add(regularizer, tf.nn.l2_loss(var))
+                regularizer = tf.add( regularizer, tf.nn.l2_loss( tf.cast(var, tf.float32) ) )
         return tf.reduce_mean(loss + beta * regularizer)
